requirements.txt

ansible == 2.1.1.0
cryptography == 1.3.4
futures >= 3.0.5
pyOpenSSL >= 0.15.1
service-identity >= 16.0.0
Twisted >= 16.3.1

